fix: Fix 3 critical model issues (Message, Payment, Review) - CRITICAL
Fixed 3 remaining critical model issues from AUDIT_COMPLET_EXHAUSTIF.md:

1. Message.php - FIXED attachment fields (line 19-22)
   ✅ Added 4 missing attachment fields to $fillable:
   - attachment (file path)
   - attachment_name (original filename)
   - attachment_type (MIME type)
   - attachment_size (file size in bytes)

   Impact: File uploads in VendorMessagesController::reply() now work
   Before: Mass assignment exception on attachment upload
   After: File attachments can be saved to messages

2. Payment.php - FIXED amount cast (line 41)
   ✅ Changed cast from 'integer' to 'float'

   Impact: Better currency handling and precision
   Before: 'integer' cast could cause issues with decimal amounts
   After: 'float' cast properly handles both cent-based and decimal storage
   Note: Maintains compatibility with existing /100 calculation for Stripe cents

3. Review.php - FIXED booking_id field (line 21, 57-60)
   ✅ Added booking_id to $fillable array
   ✅ Added booking() relationship method

   Impact: Reviews can now be linked to specific bookings
   Before: Impossible to associate review with booking
   After: Full booking->review relationship functional

All critical model issues now resolved (9/9 completed).

+++ app/Models/Payment.php
@@ -38,7 +38,7 @@ class Payment extends Model
         'paid_at' => 'datetime',
         'refunded_at' => 'datetime',
         'metadata' => 'array',
-        'amount' => 'integer',
+        'amount' => 'float',
     ];
 
     /**
